Implemented all numeric strategies. NUMERIC_BEST is now the default choice in the gui.
git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@375 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/QualityMeasure.java
@@ -491,9 +491,10 @@ public class QualityMeasure
 		itsVStructures = theDAG.determineVStructures();
 	}
 
-	public double calculateWEED(Subgroup theSubgroup)
+	public float calculateWEED(Subgroup theSubgroup)
 	{
-		return Math.pow(calculateEntropy(theSubgroup.getMembers().cardinality()),itsAlpha) * Math.pow(calculateEditDistance(theSubgroup.getDAG()),itsBeta);
+		return (float) Math.pow(calculateEntropy(theSubgroup.getMembers().cardinality()),itsAlpha) *
+			   (float) Math.pow(calculateEditDistance(theSubgroup.getDAG()),itsBeta);
 	}
 
 	public double calculateEDIT_DISTANCE(Subgroup theSubgroup)
